requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
chromadb
pypdf
python-docx
pymongo
